ZEZA story application
Its a story application created for kids in Android Studio main language used is java. This application first request sign in or sign up from the use then story list is appeared with a smooth background music. furthermore, by clicking any story user can read the story or there is an option of text to speech from which user can easily listen to the stories.